Refactoring code

+++ src/main/java/edu/asu/sbs/repositories/TransactionRepository.java
@@ -3,5 +3,5 @@ package edu.asu.sbs.repositories;
 import edu.asu.sbs.models.Transaction;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransactionRepository extends CrudRepository<Transaction,Integer> {
+public interface TransactionRepository extends CrudRepository<Transaction, Long> {
 }
